requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
dataclasses-json==0.6.2
python-dotenv==1.0.0
lxml==4.9.3
python-docx==1.1.0
flask==3.0.0
flask-cors==4.0.0
nest-asyncio==1.5.8
gunicorn==21.2.0
webdriver-manager==4.0.1
